Restructure documentation and configuration.
There was two pages for the documentation and the configuration.
Also, each html file became very long and the maintenance complex.

We restructure the documentation page to also include the configuration.
And we also split the file in several files.
The documentation and configuration file for each component is now moved to the component directory.

This commit does the following:

- Update webpack configuration to split index file and remove uikit file
- Simplify uikit components
- Style renderer now render pretty (because used in documentation file)
- Split index.html file in several files
- Split uikit.html file in several files
- Inject parts of uikit.html in index.html parts
- Remove uikit.html