Initialize `atomo-input-money` component
We continue developments by introducing a simple `atomo-input-money` component.

We take this opportunity to use `preact` as a library.
It will bring:

- Type system
- Props and state lifecycle (instead of a homemade one)

However, because of web-components,
we need to encapsulate Preact component into a homemade wrapper.
To respect types, we also need to use normalizers.
So we take this opportunity to refactor the one already made.

This commit does the following:

- Install Preact
- Use it to create the component `atomo-input-money`
- Create a homemade structure to use styles and structure components
- Refactor normalizer
- Create a wrapper preact component <-> web component