Create component `configurator`
To be able to update the configuration of the whole library but without navigating the whole page,
we create a component named `configurator`.

Its particularity is that it is at a fixed position in the page (and can be extended / collapsed).
Also, its configuration definition is in typescript instead of html.

This commit does the following:

- Create component `configurator`
- Correct component `conf-field` to be compatible with preact
- Duplicate configuration in `componentsConfiguration`